+++ x/observer/types/operational_test.go
@@ -16,6 +16,10 @@ func TestOperationalFlags_Validate(t *testing.T) {
 		errContains string
 	}{
 		{
+			name: "empty is valid",
+			of:   types.OperationalFlags{},
+		},
+		{
 			name: "invalid restart height",
 			of: types.OperationalFlags{
 				RestartHeight: -1,
@@ -49,10 +53,24 @@ func TestOperationalFlags_Validate(t *testing.T) {
 			errContains: types.ErrOperationalFlagsSignerBlockTimeOffsetLimit.Error(),
 		},
 		{
+			name: "minimum version valid",
+			of: types.OperationalFlags{
+				MinimumVersion: "v1.1.1",
+			},
+		},
+		{
+			name: "minimum version invalid",
+			of: types.OperationalFlags{
+				MinimumVersion: "asdf",
+			},
+			errContains: types.ErrOperationalFlagsInvalidMinimumVersion.Error(),
+		},
+		{
 			name: "all flags valid",
 			of: types.OperationalFlags{
 				RestartHeight:         1,
 				SignerBlockTimeOffset: ptr.Ptr(time.Second),
+				MinimumVersion:        "v1.1.1",
 			},
 		},
 	}